Inserted explicit cast in hppa version
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@1071 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/outpattern.c
@@ -611,7 +611,7 @@ void DisplayBetaStack( errorNumber, theObj, thePC)
 	theObj = (struct Object *)((unsigned)*theCell & ~1);
 	/* PC = ?????? */
 	if(theObj && isObject(theObj)) {
-	  DisplayObject(output, theObj, PC);
+	  DisplayObject(output, theObj, (long)PC);
 	} else {
 	  fprintf(output, "[Damaged object!: %x]\n", (long)theObj);
 	}
